pyproject.toml
[project]
name = "fraud-detection"
version = "0.1.0"
description = "本地端防詐騙分析系統 - 使用 RAG 和 LLM 進行詐騙案例分類"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=4.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "opencc-python-reimplemented>=0.1.7",
    "groq>=0.4.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-google-genai>=0.1.0",
    "llama-index-embeddings-huggingface>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
